requirements.txt
openai==1.12.0
twilio==8.11.0
python-dotenv==1.0.0
elevenlabs==0.2.27
pydub==0.25.1
requests==2.31.0
anthropic==0.18.1
groq==0.4.1
flask==3.0.0
flask-cors==4.0.0
pyngrok==7.0.0